requirements.txt
Django
django-cors-headers
six
django-filter
djangorestframework
djangorestframework-jwt
djoser
PyJWT
pytz
psycopg2
gunicorn
django-heroku